int to float in version no app

+++ android_app/models.py
@@ -5,7 +5,7 @@ class App(models.Model):
 	name = models.CharField(max_length=256, default='app1')
 	log = models.CharField(max_length=256)
 	link = models.URLField(blank=True)
-	version = models.IntegerField(default=1, blank=False, null=False)
+	version = models.FloatField(default=1, blank=False, null=False)
 	flag=models.BooleanField(default=False)
 	i_date = models.DateField(default= datetime.datetime(2019, 8, 31))
 	f_date = models.DateField(default= datetime.datetime(2019,9,1))
